+++ pkg/core/resources/apis/meshmultizoneservice/api/v1alpha1/helpers.go
@@ -67,7 +67,7 @@ func (m *MeshMultiZoneServiceResource) AsOutbounds() xds_types.Outbounds {
 		for _, port := range m.Spec.Ports {
 			outbounds = append(outbounds, &xds_types.Outbound{
 				Address:  vip.IP,
-				Port:     port.Port,
+				Port:     uint32(port.Port),
 				Resource: pointer.To(kri.From(m, port.GetName())),
 			})
 		}
@@ -104,7 +104,7 @@ func (p Port) GetName() string {
 	return fmt.Sprintf("%d", p.Port)
 }
 
-func (p Port) GetValue() uint32 {
+func (p Port) GetValue() int32 {
 	return p.Port
 }
